[BEKEN] add beken examples/platform files
* add beken examples/platform files
* modify beken examples/lighting-app files
