+++ examples/lighting-app/beken/BUILD.gn
@@ -15,6 +15,8 @@
 import("//build_overrides/build.gni")
 import("//build_overrides/chip.gni")
 
+examples_plat_dir = "${chip_root}/examples/platform/beken"
+
 static_library("MatterApp") {
   public_deps = [
     "${chip_root}/examples/lighting-app/lighting-common",
@@ -22,7 +24,10 @@ static_library("MatterApp") {
   ]
 
   sources = [
-    "main/CHIPDeviceManager.cpp",
+    "${examples_plat_dir}/common/BekenAppServer.cpp",
+    "${examples_plat_dir}/common/CHIPDeviceManager.cpp",
+    "${examples_plat_dir}/common/CommonDeviceCallbacks.cpp",
+    "${examples_plat_dir}/ota/OTAHelper.cpp",
     "main/DeviceCallbacks.cpp",
     "main/DsoHack.cpp",
     "main/chipinterface.cpp",
